Split integration gallery components
